+++ test/pelicanconf.py
@@ -40,6 +40,16 @@ else:
         minchin.md_it.fancy_tasklists.fancy_tasklists_plugin,
     )
 
+try:
+    import minchin.md_it.insert
+except ImportError:
+    pass
+else:
+    # add insert plugin
+    COMMONMARK["extensions"].append(
+        minchin.md_it.insert.insert_plugin,
+    )
+
 PATH = "content"
 STATIC_PATHS = [
     "images",